Allow getbalance default account with minconf=0
To get the total balance including the Unconfirmed balance the account
must be specified,  as the params are sent as positional args when
account is '' it nor the minconf will be appended into the arg list. This
allows a user to call getbalance('', 0) to get the Total (confirmed and
unconfirmed) Balance in the default account.